Hack: App settings used to specify initial route to navigate to
I have not yet found a way to specify what route to start the NativeScript
portion of the app at, but this way is working even though it feels like
a very big hack.

I am storing in app settings the route that I want to navigate too.
Getting access to a static field from the existing Java side didn't
work - my bet is that NativeScript probably did not generate a
proxy for that and only java classes with JavaScript proxies are
accessable.